### 2025-08-28 #261 @baptistehamon
LSAPy Submission
Submitting Author: @baptistehamon
All current maintainers: @baptistehamon
Package Name: LSAPy
One-Line Description of Package: Package to help and ease Land Suitability Analysis (LSA) workflow.
Repository Link:  https://github.com/baptistehamon/lsapy/
Version submitted: v0.2.0
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

LSAPy is a highly customizable Python library designed to streamline and enhance Land Suitability Analysis (LSA) workflows. The package implements a fuzzy-logic approach and provides three core objects - `SuitabilityFunction`, `SuitabilityCriteria`, and `LandSuitabilityAnalysis` that work together to deliver a flexible and user-defined LSA framework.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

LSA have been widely used by the scientific community to assess the suitability of agricultural products. In this context, LSAPy contributes to the reproducibility of such studies. Although initially intended for research purposes, LSAPy can be used by land managers/planners or for educational purposes, while being applicable to any type of land use (e.g., urban planning), thus extending its use beyond agricultural applications.

To my knowledge, the only Python package similar to LSAPy is [PyLUSAT](https://github.com/chjch/pylusat) (Python Land-Use Suitability Analysis Toolkit) but the two packages work in completely different ways. PyLUSAT provides a vector-based GIS routines and determines suitability evaluating the spatial relationship between objects, while LSAPy assesses suitability by aggregating gridded criteria indicators. Morevover, PyLUSAT workflow can easely be integrating into LSAPy providing the relevant spatial relationship between cells and objects as suitablility criteria. Finally, agricultural application of LSA often rely on gridded temporal climate data that PyLUSAT don't support, limiting its potential use for such studies.

LSAPy was pre-submitted in the issue #249.

- [x] The package is deposited in a long-term repository with the DOI: https://zenodo.org/records/15015111
