### 2025-08-30 571d1d9
Fix typo in SuitabilityCriteria repr
