### 2025-08-31 3d74396
Add conda badge and installation instructions
