### 2025-09-03 #261 @baptistehamon
Hi @eliotwrobson, thank you for the feedback! I'll fix the license file issue and improve the tutorials. Also, I forgot to mention that the paper for JOSS can be found in the `joss-submission` branch.
