### 2025-09-03 baptistehamon/lsapy#77 @baptistehamon
More comprehensive tutorials
### Addressing a Problem?

As part of the pyOpenSci [submission](https://github.com/pyOpenSci/software-submission/issues/261#issuecomment-3243751287), the EiC raised his concerns about the length of the tutorials and that, as things stand, they do not provide a comprehensive overview of LSAPy's usefulness.

### Potential Solution

Improve tutorials, better describing LSAPy's functionalities !

### Additional Context

_No response_

### Contribution

- [x] I would be willing/able to open a Pull Request to contribute this feature.

### Code of Conduct

- [x] I agree to follow this project's Code of Conduct
