### 2025-09-06 a513bfb
Update CHANGELOG.rst
