### 2025-09-06 eefd846
Update CHANGELOG.rst
