### 2025-09-07 cddee08
Rename stats_summary + small improvements
