### 2025-09-07 03eb7c2
Add example to stats_summary docstring
