### 2025-09-07 9dfacbc
Small edit in stats_summary
