### 2025-09-07 004503c
Update spatial_stats_summary and add examples doctrings
