### 2025-09-07 4570fc1
Update stats_summary type hints
