### 2025-09-07 d82834e
Fix stats_summary docstring
