### 2025-09-07 05a7b90
Add stats function to export
