### 2025-09-07 e97189e
Update CHANGELOG.rst
