### 2025-09-07 90d362d
Fix stats functions examples
