+++ src/lsapy/__init__.py
@@ -6,4 +6,4 @@ from lsapy.lsa import LandSuitabilityAnalysis
 
 __author__ = "Baptiste Hamon"
 __email__ = "baptiste.hamon@pg.canterbury.ac.nz"
-__version__ = "0.2.1-dev4"
+__version__ = "0.2.1-dev5"
