### 2025-09-11 3771e8e
Add boolean suitability function
