### 2025-09-11 be075bf
Fix boolean function message error
