### 2025-09-11 af409bf
Add boolean function tests
