### 2025-09-11 1435027
Add boolean function to doc
