### 2025-09-11 47a0a26
Update CHANGELOG.rst
