### 2025-09-14 0212e64
Make SuitabilityCriteria name and indicator not required
