### 2025-09-14 af367df
Improve setting SuitabilityCriteria attrs
