### 2025-09-14 2401ae8
Improve is_computed SuitabilityCriteria attr
