### 2025-09-14 80a0343
Add tests for SuitabilityCriteria attrs
