### 2025-09-14 2e828a0
Fix SuitabilityCriteria repr when undefined
