### 2025-09-14 35f03e7
Fix SuitabilityCriteria repr
