### 2025-10-06 4ba3b5d
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.11 → v0.13.3](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.11...v0.13.3)
- [github.com/keewis/blackdoc: v0.4.1 → v0.4.3](https://github.com/keewis/blackdoc/compare/v0.4.1...v0.4.3)
