+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
   # Use ruff as a substitute for black, flake8, isort, pyupgrade,
   # pylint.
-    rev: v0.12.11
+    rev: v0.13.3
     hooks:
       - id: ruff-check
         args: ['--fix', '--show-fixes']
@@ -68,7 +68,7 @@ repos:
       - id: mdformat
         exclude: '.github/\w+.md'
   - repo: https://github.com/keewis/blackdoc
-    rev: v0.4.1
+    rev: v0.4.3
     hooks:
       - id: blackdoc
         additional_dependencies: [ 'black==25.1.0' ]
