### 2025-10-08 2525de1
Update black version in pre-commit
