+++ .pre-commit-config.yaml
@@ -71,7 +71,7 @@ repos:
     rev: v0.4.3
     hooks:
       - id: blackdoc
-        additional_dependencies: [ 'black==25.1.0' ]
+        additional_dependencies: [ 'black==25.9.0' ]
         exclude: '.py'
   - repo: https://github.com/fluiddyn/formattex-pre-commit
     rev: 0.1.4
