+++ src/lsapy/standardize.py
@@ -1,0 +1,446 @@
+"""Standardization Functions Module."""
+
+import operator
+import warnings
+
+import numpy as np
+from scipy.optimize import curve_fit
+
+from lsapy.core.functions import declare_equation, equations, get_function_from_name
+
+try:
+    import matplotlib.pyplot as plt
+except ImportError:
+    plt = None
+    warnings.warn("Matplotlib not found. Plotting functionality will be disabled.", stacklevel=2)
+
+__all__ = ["boolean", "discrete"]
+
+FIT_TYPES = ["sigmoid", "gaussian"]  # only sigmoid and gaussian need to be fitted
+binary_ops = {">": "gt", "<": "lt", ">=": "ge", "<=": "le", "==": "eq", "!=": "ne"}
+
+
+@declare_equation("boolean")
+def boolean(x, op: str, thresh: int | float, skipna: bool = True):
+    """
+    Boolean function.
+
+    This function applies a boolean operation to the values based on a threshold.
+
+    Parameters
+    ----------
+    x : any
+        Input values.
+    op : {">", "gt", "<", "lt", ">=", "ge", "<=", "le", "==", "eq", "!=", "ne"}
+        Logical operator.
+    thresh : any
+        Threshold value.
+    skipna : bool, optional
+        Whether to skip NaN values. If True, NaN values in `x` will remain NaN in the output. Default is True.
+
+    Returns
+    -------
+    np.ndarray
+        Boolean mask of the operation.
+    """
+    if op in binary_ops:
+        op = binary_ops[op]
+    elif op in binary_ops.values():
+        pass
+    else:
+        raise ValueError(f"Operator '{op}' not recognized.")
+
+    res = getattr(operator, op)(x, thresh)
+    if skipna:
+        mask = np.isnan(x)
+        return np.where(mask, np.nan, res)
+    return res
+
+
+@declare_equation("categorical")
+@np.vectorize(otypes=[np.float32])
+def discrete(x, rules: dict[str | int, int | float]) -> np.ndarray:
+    """
+    Discrete function.
+
+    This function maps input values to a set of rules that define the output values.
+
+    Parameters
+    ----------
+    x : any
+        Input values to map.
+    rules : dict[str | int, int | float]
+        Rules to map the input values to output values. The keys correspond to the input values and the
+        values to its associated output values.
+
+    Returns
+    -------
+    np.ndarray
+        Mapped output values.
+    """
+    return rules.get(x, np.nan)
+
+
+@declare_equation("sigmoid")
+@np.vectorize(otypes=[np.float32])
+def logistic(x, a, b):
+    r"""
+    Logistic function capped to 1.
+
+    This function should be used on sigmoid-like suitability data. The function can be used on both
+    positive and negative values of `x`, as well as on increasing and decreasing sigmoid-like data.
+
+    Parameters
+    ----------
+    x : any
+        Input values.
+    a : float | int
+        Steepness of the function parameter.
+    b : float | int
+        Value of the function's midpoint.
+
+    Returns
+    -------
+    float
+        Output values.
+
+    Notes
+    -----
+    The function is defined as:
+
+    .. math::
+
+        f(x) = \frac{1}{1 + e^{-a(x - b)}}
+    """
+    return 1 / (1 + np.exp(-a * (x - b)))
+
+
+@declare_equation("sigmoid")
+def sigmoid(x):
+    r"""
+    Logistic sigmoid function.
+
+    This function is a special case of the logistic function with `a=1` and `b=0`, thus can be used on both
+    positive and negative values of `x`, however only for increasing sigmoid-like data.
+
+    Parameters
+    ----------
+    x : any
+        Input values.
+
+    Returns
+    -------
[…]
