### 2025-10-13 2f73bd3
Make functions module and SuitabilityFunction deprecated
