### 2025-10-13 8fd056e
Update SuitabilityCriteria func
