+++ tests/test_membership_fitting.py
@@ -5,39 +5,33 @@ from __future__ import annotations
 import numpy as np
 import pytest
 
-import lsapy.functions as func
-from lsapy.functions.membership import (
-    _check_fitting,  # noqa: PLC2701
-    _get_best_fit,  # noqa: PLC2701
-    _get_function_p0,  # noqa: PLC2701
-    _rmse,  # noqa: PLC2701
-    fit_membership,
-)
+import lsapy.functions.membership as mbs
+import lsapy.standardize as std
 
 
 class TestRMSE:
     def test_zero(self):
         y_true = np.array([1, 2, 3, 4, 5])
         y_pred = np.array([1, 2, 3, 4, 5])
-        assert _rmse(y_true, y_pred) == 0.0
+        assert mbs._rmse(y_true, y_pred) == 0.0
 
     def test_non_zero(self):
         y_true = np.array([1, 2, 3, 4, 5])
         y_pred = np.array([2, 3, 4, 5, 6])
-        assert _rmse(y_true, y_pred) == 1.0
+        assert mbs._rmse(y_true, y_pred) == 1.0
         y_pred = np.array([11, 12, 13, 14, 15])
-        assert _rmse(y_true, y_pred) == 10.0
+        assert mbs._rmse(y_true, y_pred) == 10.0
 
     def test_nan(self):
         y_true = np.array([1, 2, 3, 4, 5])
         y_pred = np.array([np.nan, 2, 3, 4, 5])
-        assert np.isnan(_rmse(y_true, y_pred))
+        assert np.isnan(mbs._rmse(y_true, y_pred))
 
 
 class TestCheckFitting:
     def test_all(self):
         # should skip sigmoid and vetharaniam2024_eq8
-        valid, skipped = _check_fitting()
+        valid, skipped = mbs._check_fitting()
         assert valid == [
             "logistic",
             "vetharaniam2022_eq3",
@@ -47,36 +41,36 @@ class TestCheckFitting:
         assert skipped == ["sigmoid", "vetharaniam2024_eq8"]
 
     def test_type_like(self):
-        valid, skipped = _check_fitting("sigmoid_like")
+        valid, skipped = mbs._check_fitting("sigmoid_like")
         assert valid == ["logistic", "vetharaniam2022_eq3", "vetharaniam2022_eq5"]
         assert skipped == ["sigmoid"]
 
     def test_list(self):
         # should not str type in list
-        valid, skipped = _check_fitting(["logistic", 1])
+        valid, skipped = mbs._check_fitting(["logistic", 1])
         assert valid == ["logistic"]
         assert skipped == []
 
     def test_invalid(self):
         # test for invalid type
         with pytest.raises(ValueError, match="`fit_on` should be a str or a list of string. Got <class 'int'>"):
-            _check_fitting(1)
+            mbs._check_fitting(1)
         # test for invalid name
         with pytest.raises(ValueError, match="No functions to fit. Try to modify `fit_on` parameter."):
-            _check_fitting("invalid")
+            mbs._check_fitting("invalid")
 
 
 class TestGetFunctionP0:
     def test_sigmoid(self):
-        p0 = _get_function_p0("vetharaniam2022_eq5", np.array([1, 2, 3, 4, 5]))
+        p0 = mbs._get_function_p0("vetharaniam2022_eq5", np.array([1, 2, 3, 4, 5]))
         assert p0 == [1, 3]
 
     def test_gaussian(self):
-        p0 = _get_function_p0("vetharaniam2024_eq10", np.array([1, 2, 3, 4, 5]))
[…]
