### 2025-10-16 1e2c852
Add standardization functions tests
