### 2025-10-16 dc78c0b
Update docstrings examples
