### 2025-10-16 2504d5d
Bump version: 0.2.1-dev7 → 0.2.1-dev8
Signed-off-by: Baptiste Hamon <baptiste.hamon@pg.canterbury.ac.nz>
