### 2025-11-03 c38a210
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.13.3 → v0.14.3](https://github.com/astral-sh/ruff-pre-commit/compare/v0.13.3...v0.14.3)
- [github.com/executablebooks/mdformat: 0.7.22 → 1.0.0](https://github.com/executablebooks/mdformat/compare/0.7.22...1.0.0)
- [github.com/keewis/blackdoc: v0.4.3 → v0.4.5](https://github.com/keewis/blackdoc/compare/v0.4.3...v0.4.5)
