+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
   # Use ruff as a substitute for black, flake8, isort, pyupgrade,
   # pylint.
-    rev: v0.13.3
+    rev: v0.14.3
     hooks:
       - id: ruff-check
         args: ['--fix', '--show-fixes']
@@ -63,12 +63,12 @@ repos:
       - id: rst-inline-touching-normal
       - id: text-unicode-replacement-char
   - repo: https://github.com/executablebooks/mdformat
-    rev: 0.7.22
+    rev: 1.0.0
     hooks:
       - id: mdformat
         exclude: '.github/\w+.md'
   - repo: https://github.com/keewis/blackdoc
-    rev: v0.4.3
+    rev: v0.4.5
     hooks:
       - id: blackdoc
         additional_dependencies: [ 'black==25.9.0' ]
