### 2025-11-03 4fd1aca
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
