### 2025-11-04 3c18719
Fix SuitabilityFunctions section order
