### 2025-11-05 7aaa92b
Force Python 3.13 max in CI
