+++ .github/workflows/publish-pypi.yml
@@ -15,6 +15,9 @@ jobs:
     permissions:
       id-token: write # IMPORTANT: mandatory for trusted publishing
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ['3.13']
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
@@ -23,7 +26,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: '3.x'
+          python-version: ${{ matrix.python-version }}
       - name: Install CI libraries
         run: |
           python -m pip install --require-hashes -r ci/requirements_ci.txt
