### 2025-11-06 928e837
Make SuitabilityCriteria name an attr
