### 2025-11-06 3d2f035
Update CHANGELOG.rst
