### 2025-11-07 3da2153
Update CHANGELOG.rst
