### 2025-11-07 64d1a9b
Fix CHANGELOG.rst link issue
