### 2025-11-07 02af0fc
Fix stats function docstring issue
