### 2025-11-07 49d2d72
Update CHANGELOG.rst
