### 2025-11-09 fa8babf
Bump version: 0.2.1-dev12 → 0.3.0-dev0
Signed-off-by: Baptiste Hamon <baptiste.hamon@pg.canterbury.ac.nz>
