### 2025-11-09 6f08cfe
Bump version: 0.3.0-dev0 → 0.3.0
Signed-off-by: Baptiste Hamon <baptiste.hamon@pg.canterbury.ac.nz>
