### 2025-11-24 807e059
Remove functions module
