+++ docs/paper/paper.bib
@@ -154,6 +154,16 @@
 }
 
 
+@article{hamon2025e,
+  title = {{NZLUSDB}: New Zealand Land Use Suitability Database},
+  url = {https://github.com/baptistehamon/nzlusdb},
+  journaltitle = {[Manuscript in preparation]},
+  author = {Hamon, Baptiste and Mazzilli, Gabriel and Quénol, Hervé and Vannier, Clémence
+    and Cochrane, Thomas A.},
+  date = {2025}
+}
+
+
 @article{hoyer2017,
   title = {xarray: N-D labeled Arrays and Datasets in Python},
   volume = {5},
@@ -203,6 +213,24 @@
   author = {Kalogirou, S},
   urldate = {2025-08-24},
   date = {2002-03},
+  langid = {english}
+}
+
+
+@article{malczewski2004,
+  title = {{GIS}-based land-use suitability analysis: a critical overview},
+  volume = {62},
+  issn = {03059006},
+  url = {https://linkinghub.elsevier.com/retrieve/pii/S0305900603000801},
+  doi = {10.1016/j.progress.2003.09.002},
+  shorttitle = {{GIS}-based land-use suitability analysis},
+  pages = {3--65},
+  number = {1},
+  journaltitle = {Progress in Planning},
+  shortjournal = {Progress in Planning},
+  author = {Malczewski, Jacek},
+  urldate = {2025-10-27},
+  date = {2004-07},
   langid = {english}
 }
 
