### 2025-12-01 d2b12c7
Fix paper typo
