### 2025-12-01 a3a05d3
Update paper typo
