### 2025-12-19 0190da4
Update fig2 location in paper
