+++ docs/paper/paper.md
@@ -58,9 +58,9 @@ Finally, the `compute()` method applies the standardization function to the give
 ## Example Workflow
 For this example, a land suitability analysis was performed using the soil and climate data provided in LSAPy, covering the New Zealand territory. Three suitability criteria were defined: slope, annual precipitation excess, and the growing temperature requirement. The first criteria was categorised under "soilTerrain", while the latter two were grouped under "climate". The indicators associated with each criteria were the slope (in degrees), annual total precipitation (in mm), and annual mean temperature (in °C) respectively, and are presented in (\autoref{fig:example}). It is important to note that the chosen criteria, indicators and standardisation functions are for demonstration purposes only and do not correspond to an actual land suitability analysis for any specific crop. A comprehensive LSA would typically involve a larger number and more complex criteria.
 
-![Workflow of the land suitability analysis example. Presented are the input indicators associated with the criteria used, the standardisation functions applied and the output suitability. The suitability of each of the criteria is shown, as well as the aggregated suitability of the categories and the final land suitability.\label{fig:example}](example.png)
-
 The suitability of each criteria was computed by applying the defined standardisation functions ((\autoref{fig:example})). The criteria suitability were then aggregated: first, the suitability of the "soilTerrain" and "climate" categories were calculated, and then the overall land suitability was derived by combining the two categories. In this simple example, the arithmetic mean was used for all aggregation steps. The resulting land suitability map is also presented in (\autoref{fig:example}) and allows identifying the most suitable areas based on the defined criteria.
+
+![Workflow of the land suitability analysis example. Presented are the input indicators associated with the criteria used, the standardisation functions applied and the output suitability. The suitability of each of the criteria is shown, as well as the aggregated suitability of the categories and the final land suitability.\label{fig:example}](example.png)
 
 # Research Applications
 
