### 2025-12-19 577bf09
Fix paper typo
