### 2026-01-05 3b22633
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.14.7 → v0.14.10](https://github.com/astral-sh/ruff-pre-commit/compare/v0.14.7...v0.14.10)
- [github.com/numpy/numpydoc: v1.9.0 → v1.10.0](https://github.com/numpy/numpydoc/compare/v1.9.0...v1.10.0)
