+++ .pre-commit-config.yaml
@@ -25,7 +25,7 @@ repos:
   - repo: https://github.com/astral-sh/ruff-pre-commit
   # Use ruff as a substitute for black, flake8, isort, pyupgrade,
   # pylint.
-    rev: v0.14.7
+    rev: v0.14.10
     hooks:
       - id: ruff-check
         args: ['--fix', '--show-fixes']
@@ -37,7 +37,7 @@ repos:
         additional_dependencies: [ 'tomli']
         args: [ '--toml=pyproject.toml']
   - repo: https://github.com/numpy/numpydoc
-    rev: v1.9.0
+    rev: v1.10.0
     hooks:
       - id: numpydoc-validation
         exclude: ^docs/|^tests/|noxfile.py|^src/lsapy/core/formatting.py
