### 2026-01-05 e2f58ff
Add HDF5 installation on Linux in CI
