+++ .github/workflows/testing.yml
@@ -46,6 +46,11 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+      - name: Install HDF5 (Linux)
+        if: matrix.os == 'ubuntu-latest'
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y libhdf5-dev
       - name: Install CI libraries
         run: |
           python -m pip install --require-hashes -r ci/requirements_ci.txt
