### 2026-01-05 56418a8
Update CHANGELOG.rst
