### 2026-02-06 92f50ee
Add h5py to dev dependencies
