+++ pyproject.toml
@@ -47,6 +47,7 @@ dev = [
   "coverage[toml] >=7.7.0",
   "flit >=3.11",
   "h5netcdf >=1.4.0",
+  "h5py",
   "ipython >=8.20.0",
   "matplotlib >=3.9.0",
   "mypy >=1.17.1",
