### 2026-02-20 4b89417
Fix sc_repr when empty xr.DataArray indicator
