+++ src/lsapy/__init__.py
@@ -5,4 +5,4 @@ from lsapy.lsa import LandSuitabilityAnalysis
 
 __author__ = "Baptiste Hamon"
 __email__ = "baptiste.hamon@pg.canterbury.ac.nz"
-__version__ = "0.3.2-dev0"
+__version__ = "0.3.2-dev1"
