### 2026-04-25 #261 @tkoyama010
## Editor Response

:wave: Hi @baptistehamon! I apologize for the delay — I am @tkoyama010 and I will be serving as the editor for your LSAPy submission. I am looking forward to working with you on this!

I have conducted the following editor checks:

### Editor Checks

- [x] **Fit**: The package meets pyOpenSci's [scope criteria](https://www.pyopensci.org/software-peer-review/about/package-scope.html) in the data processing/munging and geospatial categories.
- [x] **Automated tests**: Package has a test suite and CI via GitHub Actions.
- [x] **License**: The package has an [OSI approved license](https://opensource.org/licenses) (MIT).
- [x] **Repository**: The repository link resolves correctly.
- [x] **Package overlap**: LSAPy's raster/gridded approach is sufficiently distinct from PyLUSAT's vector-based GIS approach.
- [x] **Archive**: Zenodo DOI resolves correctly (https://zenodo.org/records/15015111).
- [x] **Version**: v0.2.0 matches the GitHub release.

I will serve as **Reviewer 1** for this submission. I am currently searching for a **Reviewer 2** and will keep you updated here. Thank you for your patience!
