### 2026-04-25 #261 @tkoyama010
*Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics.*

- [x] As the reviewer, I confirm there are no conflicts of interest for me to review this work (consult your editor if unsure).

- [x] A statement of need clearly stating problems the software solves and its target audience in the README
- [x] Installation instructions for the development version and non-standard dependencies in README
- [x] Short quickstart tutorials demonstrating significant functionality that runs locally
- [x] Function Documentation for all user-facing functions
- [x] Examples for all user-facing functions
- [x] Community guidelines including contribution guidelines in README or CONTRIBUTING
- [x] Metadata including author(s), author e-mail(s), a URL, and other relevant metadata (e.g., `pyproject.toml`)

### README Requirements

- [x] Package has a README file in the root directory
- [x] Badges for continuous integration and test coverage
- [x] Badges for docs building (if applicable)
- [x] Badges for Python versions supported
- [x] Badges for current package version (on PyPI / Conda)
- [x] Short description of package goals
- [x] Any additional setup required (authentication tokens, etc.)
- [x] Descriptive links to all vignettes
- [x] Brief demonstration of package usage
- [x] Link to documentation website
- [ ] How the package compares to similar packages (if applicable)

### Usability

- [x] Package documentation is clear and easy to find and use
- [x] All functions have documentation and associated examples

- [x] Installation succeeds as documented
- [x] Functional claims of the software have been confirmed
- [x] Performance claims of the software have been confirmed
- [x] All tests pass on reviewer's local machine for the submitted version
- [x] Tests cover essential functions and a reasonable range of inputs and conditions
- [x] Continuous integration setup exists (GitHub Actions or other acceptable CI platform)
- [x] Package conforms to pyOpenSci packaging guidelines
- [x] Package supports modern versions of Python (not End of Life versions)
- [x] Code format follows PEP 8 guidelines (CI tests for linting pass)

### For Packages Submitting to JOSS

- [x] Package has an obvious research application per JOSS's definition
- [x] Package contains a `paper.md` with a short summary, authors with affiliations, a statement of need, and references with DOIs

### Final Approval (Post-Review)

- [ ] The author responded to my review and made changes to my satisfaction. I recommend approving this package.

**Estimated hours spent reviewing:** 3

---

## Review Comments

Thank you for submitting LSAPy! All three tutorials ran without errors and the documentation is well-organized. Please address the following before approval.

### Required Changes

1. **`matplotlib` missing from required dependencies** — `import lsapy` fails after a clean `pip install lsapy` with `ModuleNotFoundError: No module named 'matplotlib'`.
2. **`h5py` missing from dependencies** — `test_utils.py::TestOpenData::test_open_land` fails with `ImportError: No module named 'h5py'`.
3. **`lsapy.standardize` not importable as a submodule** — `from lsapy.standardize import logistic` raises `ModuleNotFoundError`. Please expose the submodule or update the documentation to use `from lsapy import logistic`.

### Minor Suggestions

- Add a citation/BibTeX section to the README.
- Add a link to development installation instructions on the Installation page.
