### 2026-05-18 #261 @tkoyama010
Hi @baptistehamon, I have submitted a pull request to add `h5netcdf` as a required dependency, which should resolve Issue 2: https://github.com/baptistehamon/lsapy/pull/153
